+++ src/Conf/Conf.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Scriptotek\Alma\Conf;
+
+use Scriptotek\Alma\Client;
+
+class Conf
+{
+    public function __construct(Client $client)
+    {
+        $this->client = $client;
+        $this->libraries = new Libraries($client);
+    }
+}